backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
gunicorn==20.1.0  # Для продакшена